extensions/itx-gates/requirements.txt
PyYAML==6.0.2
typing_extensions==4.15.0
